rustlings - conversions

+++ exercises/23_conversions/using_as.rs
@@ -8,13 +8,10 @@
 // returns the proper type.
 //
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a
-// hint.
-
-// I AM NOT DONE
-
+// hint.
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    total / (values.len() as f64)
 }
 
 fn main() {
